(PE-24734) Fix puppet infra status collection
This commit fixes the `puppet infrastructure status` call added in f3e8a93 by
adding a guard to ensure the `/opt/puppetlabs/bin/puppet-infrastructure`
command exists before running the check. The dependency on the
`pe-client-tools` package is also dropped as it has nothing to do with
providing the `puppet infrastructure` command.