(PE-24757) Update directory setting collection
This commit updates the RBAC check to fetch directory settings from the RBAC
API instead of going through the database. This change allows the collection
to stay current when new settings are added and also allows the check to
function if run on a node where Console Services is present, but the database
is not.