(PE-24757) Load heredocs using cat instead of read
This commit swaps the use of `read -r -d '' <var name>` for
`<var name>=$(cat ...)` as read will exit 1 when it hits EOF without finding
the delimiter set by `-d ''`. This is fine in normal use, but causes tests
to fail as those run under `set -e` which causes the script to exit with an
error when read exits 1.