(PE-15844) Remove unnecessary command substitution
This commit removes some unnecessary command substitutions in if statements
that was flagged by shell check. The use of command substitution meant that
the if statement was testing the result of executing the command's output
instead of the command its self. All commands being tested had their output
streams redirected to /dev/null and ended up working as intended. However,
it is still a dangerous pattern.