(PE-15844) Separate variable declarations
This commit separates variable declaration from value assignment when the value
is produced by executing a command. In these cases, declaring and assigning at
the same time masks the exit code of the command. This bit of lint was actually
hiding some calls to `puppet master --configprint` that fail on PE 2019.0.x.