(PE-25578) Add cgroup data to support script
Prior to this commit, we did not pull any cgroup specific data for
the PE services.  This left a support gap when customers had limits
in place that caused services to fail.

With this commit, we copy over the /sys/fs/cgroup directories for
each PE service and place them under the system directory.