(PE-26173) Avoid orchestrator encryption keys
This commit tightens the configuration file whitelists in the Ruby and Bash
versions of the support script to explicitly list which orchestration
configuration files to capture instead of copying the entire conf.d directory.
In PE 2019.1, the Orchestrator started storing encryption keys under this
directory that should not be included in support script output.