(PE-26141) Implement base Check class for diagnostics
This commit adds a Check class to the Ruby support script. Diagnostics may be
implemented as subclasses that provide a `run` method that executes the
diagnostic and an optional `setup` method that configures any logics or
constraints that apply to the diagnostic. The Confinable module is also
included so that the `setup` method may use `confine` statements to restrict
when a diagnostic should execute.