(PE-26141) Confinable objects may be conditionally enabled
This commit adds a new `enabled?` flag to the Confinable module that can be
used to conditionally enable or disable confined objects. This can be used
by Scope or Check classes for which a given operation is suitable, i.e.
the operating system is Windows, but that still should not run by default
unless explicitly requested (gathers lots of data, gathers sensitive data,
etc.)