(PE-26141) Add enable, disable, and only settings
This commit adds a top-level Settings object that can be configured with
`enable`, `disable`, and `only` lists which contain the names of Scopes
or Checks. The behaviors are:

  - only: Only the listed items are enabled. All other items are disabled.
          Items which default to disabled stay disabled unless explicitly
          listed in `only`.

  - enable: The listed items are enabled. This can be used to enable
            items which default to being disabled.

  - disable: The listed items are disabled.

The order of precedence is as listed above. This means that the scope of
enabled items can be truncated with `only`, broadened with `enable` and
finally reduced with `disable`.