(PE-25578) Ensure only readable cgroup files are copied
The "files" under the cgroup mount are a bit special as they are not entries
on disk, but rather a representation of state held in the kernel. Some are
not readable, which causes `cp` to fail. This commit updates the collection
use `find -readable` to filter out any write-only entries.