(PE-26142) Extract script options to Settings class
This commit moves options that influence the behavior of the Ruby support
script from the Support class to the SupportScript::Settings class. Defaults
set by the CLI parser along with validation code from Support are also
centralized in Settings.

A new SupportScript::Runner class is also introduced which takes care of
initializing and executing script components.