Sync Ruby script CLI defaults with support face
The Ruby support script is commonly invoked through the Puppet Face system
via `puppet enterprise support`. This passes default values for the following
options:

  - dir
  - log_age
  - scope
  - ticket

All other options are left unset. In contrast, the CLI built into the Ruby
script sets a default value for each option. Some of these defaults, such
as empty strings, cause failures when the script is run directly.

This commit syncs the behaviors of the Puppet Face and CLI by removing all
defaults from the CLI that are not also set by the Face.