Add console logger to Ruby support script
This commit adds a `LogManager.console_logger` method that returns a standard
Ruby Logger instance pre-configured to send messages to stderr with a simple
format. The CLI code is updated to add this to the main LogManager instance
before dispatching to the Runner for execution.

stderr is used so that eventually the path of the support archive created will
be the only thing sent to stdout, which will facilitate chaining the support
script with other tools.