(FM-7833) adding in referece to the ace-server for kearney
The ACE server will be a new service added in Kearney. The service is similar to bolt-server, as they are both Puma services so the example set by bolt-server should be valid and the same diagnositics should be gathered.

The service will be added in Kearney, but I believe it won't be in use until a z release.