(PE-26142) Add base scope and check
This commit adds a base scope that contains all other diagnostics and
fixes a bug that prevented scopes from properly enabling if the `only`
and `enabled` settings are empty. The DiagnosticHelpers module is
also included in the base Check and Scope classes.

A BaseStatus check is also added which writes out the metadata.json
file.