(PE-26142) Lift Puppet agent diagnostics to a check
This commit moves the diagnostics related to the Puppet agent to
Check::PuppetAgentStatus which inherits from Check::ServiceStatus.

The new check gathers:

  - Facter output and debug-level logs
  - A list of gems installed in the Puppet Ruby environment
  - Whether Puppet's configured server hostname responds to a ping
  - A copy of classes.txt, graphs/, last_run_summary.yaml, and
    resources.txt from Puppet's statedir.

This commit also updates the DiagnosticHelpers module with a puppet_conf
function that retrieves the value of Puppet settings and enhances the
exec_drop function with the ability to send stderr to a separate file.