(PE-27863) Allow diagnostics to set an exit code
This commit adds an `exit_code` field to the the `state` object that
the `Runner#run` method will use as a return value. This field is
initialized to 0, but can be used to flag a run as failed without
needing to halt execution or thread an error all the way through
the Scope and Check instances.