(PE-27863) Switch to os.family for pkg detection
This commit updates the logic for selecting the OS package manager to
use the `os.family` fact instead of `os.name`. This allows the support
script to run properly on platforms in the RedHat family such as
Oracle Linux and Scientific Linux.

The exit code of the script is also set to 1 if no suitable package
manager can be selected.