(PE-27890) Use libc handlers for INT, TERM, and QUIT signals
This commit updates the `Runner#setup` routine of the Ruby Support
Script to configure SIGINT, SIGTERM, and SIGQUIT to use the default
handlers provided by libc instead of the Ruby runtime. This is needed
because the script currently uses the `Facter::Core::Execution` API
to run external commands and that library waits for commands to finish
inside a C++ routine while holding Ruby's Global VM Lock (GVL). Thus, the
default Ruby signal handlers cannot run until the lock is released.
This behavior around the GVL results in the support script ignoring
`CTRL-C`, `CTRL-\`, or a plain `kill` command.

Using the libc handlers means we loose the ability to send messages to
the user or perform cleanup actions in Ruby code. However, we effectively
don't have this ability with the current Facter behavior. Maintaining
responsiveness to user cancelation requests is of critical importance
in enterprise environments as opening a root shell can be an intensive
process that requires clearing change requests and other amounts of red tape.
The support script cannot stall a root shell that has been allocated for
debugging when it has no easy option for cancelation.

Ref. https://tickets.puppetlabs.com/browse/FACT-2250