(PE-16083) Create Bundler group for acceptance test deps
This commit moves Beaker and friends to an "acceptance" group inside the
Gemfile. Some of the Beaker plugins are only available with a connection
to the Artifactory instance on the Puppet corporate network. Segmenting
them allows these items to be skipped when provisioning publicly hosted
test runners to execute unit tests.