(PE-16083) Fix bundle install
This commit corrects the `--with-out` flag to `--without` (thanks
for adding "smart hyphenation" to `--help` output Bundler).

Declaration of the "beaker-pe-large-environments" dependency is also
skipped in the Gemfile if the GEM_SKIP_INTERNAL environment variable
is set. This is because Bundler queries the `source` for info on
every `gem` dependency --- even those that are going to be skipped
by `--without`.